move main secure note tests
- test_item_types/test_secure_note_item.py
